+++ Cargo.toml
@@ -10,3 +10,15 @@ crate-type = ["cdylib"]
 
 [dependencies]
 wasm-bindgen = "0.2.63"
+
+[profile.test]
+opt-level = 3
+debug = false
+split-debuginfo = '...'  # Platform-specific.
+debug-assertions = false
+overflow-checks = false
+lto = true
+panic = 'unwind'
+incremental = false
+codegen-units = 1
+rpath = false
