+++ src/algorithms.rs
@@ -1,0 +1,40 @@
+#[allow(clippy::too_many_arguments)]
+fn rdp_step(x: &[f32], y: &[f32], first: usize, last: usize, epsilon: f32, xnew: &mut Vec<f32>, ynew: &mut Vec<f32>, count: &mut f32) {
+    let mut dmax: f32 = 0.0;
+    let mut idx: usize = 0;
+    let mut d: f32;
+    let m = (y[first] - y[last]) / (x[first] - x[last]);
+    let c = y[first] - m * x[first];
+    let modulus = (1.0 + m.powi(2)).sqrt();
+    for i in (first + 1)..last {
+        d = (y[i] - m * x[i] - c).abs() / modulus;
+        if d > dmax {
+            idx = i;
+            dmax = d;
+        }
+    }
+    if dmax > epsilon {
+        if idx > first + 1 {
+            rdp_step(x, y, first, idx, epsilon, xnew, ynew, count)
+        } 
+        *count += 1.0;
+        xnew.push(x[idx]);
+        ynew.push(y[idx]);
+        if last > idx + 1 {
+            rdp_step(x, y, idx, last, epsilon, xnew, ynew, count)
+        } 
+    }
+}
+
+pub fn ramer_douglas_peucker(x: &[f32], y: &[f32], epsilon: f32) -> [Vec<f32>; 2] {
+    let last = x.len() - 1;
+    let mut xnew = Vec::with_capacity(20);
+    let mut ynew = Vec::with_capacity(20);
+    xnew.push(x[0]);
+    ynew.push(y[0]);
+    let mut count: f32 = 2.0;
+    rdp_step(x, y, 0, last, epsilon, &mut xnew, &mut ynew, &mut count);
+    xnew.push(x[110]);
+    ynew.push(y[110]);
+    [xnew, ynew]
+}
